Firmware update implementation (#259)
* added a new firmwate to manage firmware update

Implemented the OB register used by the IOT module during firmware update.
created a new .c and .h file to handle firmware update procedure.
modified the name of some defines and variables.

* replaced task function for normal function

Th idea is use function to be called from the timer interruption.
Doing this, the response to the firmware update will be more fast.
Timer interruption is 0.5 ms.

* firmware read to test

ON this commit device can receive data frames from IOT module using the protocol implemented by IOT.

* fixed bug that could make the system lost the first data frame

* fixed a problem on error state

* fixing some problems and changed setup to improve communication

* changed configuration on SDO memory

* changed canbus to rxfifo with interruption on reception

* fixed timeout bug on firmware upload and improved canbus

* added a new subindex to inform data frame number

timer used by the function had interruption priority changed from 12 to 4.
CANBUS reception and tx priority were changed from 4 to 7.

This change is necessary to give enough time to the firmware update function read and process all received information.

* fixed a problem with command update when finished update

* created and start little_fs integration

* integration

* little fs working with firmware update

* added crc check to received payload and data payload in the external memory

* added crc but update  stops after the 4 data frame

* using task to receive firmware update

We have a race condition between timer, spi and CAN.

* nor wokriing

Aaaaa

* possible fix to the firmware update

* detected the stuck problem - spi is blocking for 6 seconds

* memory must be erased to work .

function fs_erase is not ok to erase the file. just a erase all(drive function) is capable to erase the memory.
we still have a peoblem at the end of the file.
read and write process to verify the received firmware take much time, so sometimes frames are lost.

* removed SF_Storage_Erase function and added Serial_Flash_EraseChip.

* added read function and read or write.

* turned off optimization of the serial flash file

* fixed some issues

* now only writing on the external flash

switch read and write break the externa memory format.
only write and read on sequency works.

* fixed payload detection bug

* add python file to support .bin file

* created files from fsp 3.7.0

Aligned FSP CAN, SPI, TIMER setup.

* update bootloader and scat region info

* bootloader 1 fix

* fixed mcuboot address.

* small adjustments

deleted some comments and not used lines.
using uart 1 to screen communication.

* changed FPS from 3.8.0 to 3.7.0

changed memory config from 512k to 256k

* fixed OD sub index OTA command read size.

* update some comments.

* added some ASSERT and deleted some unless comments

* PR requested modification

* Create README.md

* fixed indentation and canlog task priority to low

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/src/middleware_vehicle/power_enable.c
@@ -108,7 +108,7 @@ void PWREN_StopPower(PWREN_Handle_t * pHandle)
          address responsible to hold the firmware update command.
   @return void
 */
-void PWREN_CheckFirmwareUpdateCommand(PWREN_Handle_t * pHandle, uint16_t UpdateCommand)
+void PWREN_CheckFirmwareUpdateCommand(PWREN_Handle_t * pHandle, uint8_t UpdateCommand)
 {
     
     ASSERT(pHandle != NULL);
